Limit numer of child rays
The ray_color function is recursive. It will only stop recursing when it
fails to hit anything. In some cases, this will take a long time -- long
enough to blow the stack. To guard against that, we limit the maximum
recursion depth, returning no light contribution at the max recursion
depth.