Add function to check if vector is near zero
If random unit vectors we generate is exactly opposite the normal
vector, the two will sum to zero. This results in a zero scatter
direction vector. This will lead to bad scenarios, thus we need to
intercept the condition before we pass it on.

The `near_zero` function will return true if the vector is very close to
zero in all dimensions.